// stock-manager-client/src/app/api/model/produitDTO.ts
/**
 * API Gestion Produits
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategorieDTO } from './categorieDTO';


export interface ProduitDTO { 
    id?: number;
    nom: string;
    prix: number;
    categorieId: number;
    categorie?: CategorieDTO;
    image?: string;
}