// stock-manager-client/src/app/api/model/produitUpdateDTO.ts
/**
 * API Gestion Produits
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProduitUpdateDTO { 
    nom?: string;
    prix?: number;
    categorieId?: number;
    image?: string;
}